+++ src/web/js/popup.js
@@ -9,18 +9,28 @@ const selectThemeElName = '#js-select-theme input:radio[name="theme"]'
 const selectFontElName = '#js-select-font'
 const manifestData = window.zadark.browser.getManifest()
 
+const switchHideLatestMessageElName = '#js-switch-hide-latest-message'
+const switchBlockTypingElName = '#js-switch-block-typing'
+const switchBlockSeenElName = '#js-switch-block-seen'
+const switchBlockDeliveredElName = '#js-switch-block-delivered'
+
 const MSG_ACTIONS = {
+  CHANGE_THEME: '@ZaDark:CHANGE_THEME',
+  CHANGE_FONT: '@ZaDark:CHANGE_FONT',
+  CHANGE_HIDE_LATEST_MESSAGE: '@ZaDark:CHANGE_HIDE_LATEST_MESSAGE',
   GET_ENABLED_BLOCKING_RULE_IDS: '@ZaDark:GET_ENABLED_BLOCKING_RULE_IDS',
   UPDATE_ENABLED_BLOCKING_RULE_IDS: '@ZaDark:UPDATE_ENABLED_BLOCKING_RULE_IDS'
 }
 
 $(versionElName).html(`Phiên bản ${manifestData.version}`)
 
+// Init popup theme
 window.zadark.utils.refreshPageTheme()
-window.zadark.utils.refreshPageFont()
-window.zadark.browser.getExtensionSettings().then(({ theme, font }) => {
+
+window.zadark.browser.getExtensionSettings().then(({ theme, font, enabledHideLatestMessage }) => {
   $(selectThemeElName).filter(`[value="${theme}"]`).attr('checked', true)
   $(selectFontElName).val(font)
+  $(switchHideLatestMessageElName).prop('checked', enabledHideLatestMessage)
 })
 
 $(selectThemeElName).on('change', async function () {
@@ -31,8 +41,8 @@ $(selectThemeElName).on('change', async function () {
   // Set popup theme
   window.zadark.utils.refreshPageTheme()
 
-  // Set page theme for all Zalo tabs
-  window.zadark.browser.sendMessage2ZaloTabs('@ZaDark:CHANGE_THEME', { theme })
+  // Set theme for all Zalo tabs
+  window.zadark.browser.sendMessage2ZaloTabs(MSG_ACTIONS.CHANGE_THEME, { theme })
 })
 
 $(selectFontElName).on('change', async function () {
@@ -40,8 +50,16 @@ $(selectFontElName).on('change', async function () {
 
   await window.zadark.browser.saveExtensionSettings({ font })
 
-  // Set page font for all Zalo tabs
-  window.zadark.browser.sendMessage2ZaloTabs('@ZaDark:CHANGE_FONT', { font })
+  // Set font for all Zalo tabs
+  window.zadark.browser.sendMessage2ZaloTabs(MSG_ACTIONS.CHANGE_FONT, { font })
+})
+
+$(switchHideLatestMessageElName).on('change', async function () {
+  const enabledHideLatestMessage = $(this).is(':checked')
+  await window.zadark.browser.saveExtensionSettings({ enabledHideLatestMessage })
+
+  // Set enabledHideLatestMessage for all Zalo tabs
+  window.zadark.browser.sendMessage2ZaloTabs(MSG_ACTIONS.CHANGE_HIDE_LATEST_MESSAGE, { enabledHideLatestMessage })
 })
 
 const handleBlockingRuleChange = (elName, ruleId) => {
@@ -56,22 +74,7 @@ const handleBlockingRuleChange = (elName, ruleId) => {
   }
 }
 
-const initPrivacy = () => {
-  const isSupportPrivacy = window.zadark.utils.getIsSupportPrivacy()
-
-  const panelPrivacyElName = '#js-panel-privacy'
-  const panelPrivacyNotAvailableElName = '#js-privacy-not-available'
-
-  const switchBlockTypingElName = '#js-switch-block-typing'
-  const switchBlockSeenElName = '#js-switch-block-seen'
-  const switchBlockDeliveredElName = '#js-switch-block-delivered'
-
-  if (!isSupportPrivacy) {
-    $(panelPrivacyNotAvailableElName).html(`Chưa hỗ trợ trên ${window.zadark.browser.name}`)
-    $(panelPrivacyElName).addClass('not-available')
-    return
-  }
-
+const enableBlocking = () => {
   chrome.runtime.sendMessage({ action: MSG_ACTIONS.GET_ENABLED_BLOCKING_RULE_IDS }).then((ruleIds) => {
     if (!Array.isArray(ruleIds)) {
       return
@@ -87,4 +90,22 @@ const initPrivacy = () => {
   $(switchBlockDeliveredElName).on('change', handleBlockingRuleChange(switchBlockDeliveredElName, 'rules_block_delivered'))
 }
 
-initPrivacy()
+const disableBlocking = () => {
+  const disabledList = [switchBlockTypingElName, switchBlockSeenElName, switchBlockDeliveredElName]
+
+  disabledList.forEach((elName) => {
+    $(elName).parent().parent().addClass('zadark-switch__disabled')
+  })
+}
+
+const initBlocking = () => {
+  const isSupportBlocking = window.zadark.utils.getIsSupportBlocking()
+
+  if (isSupportBlocking) {
+    enableBlocking()
+  } else {
+    disableBlocking()
+  }
+}
+
+initBlocking()